% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/drop_percent_missing.R
\name{drop_percent_missing}
\alias{drop_percent_missing}
\title{Drop columns with over a certain percent missing.}
\arguments{
\item{x}{any. The column to screen.}

\item{threshold}{numeric. The percent (e.g., 0.5) below which to remove
columns that have at least that percent missing. The default is 0.8.}

\item{missing_level}{character. A string representing a missing
value in a factor. This will only be used when processing
character and factor columns.}
}
\description{
Drop columns with over a certain percent missing.
}

