% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.syberiaMungebits2.R
\name{is.mungebit_generator}
\alias{is.mungebit_generator}
\title{Whether or not a function is a mungebit generator.}
\usage{
is.mungebit_generator(obj)
}
\arguments{
\item{obj}{ANY. An R object to test.}
}
\value{
TRUE or FALSE according as \code{obj} has the
  class \code{"mungebit_generator"}.
}
\description{
Mungebit generators are arity-0 functions that generate
a new mungebit when called.
}
