% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.syberiaMungebits2.R
\name{mungebit_generators}
\alias{mungebit_generators}
\title{All mungebit generators exported by the syberiaMungebits2 package.}
\usage{
mungebit_generators()
}
\value{
A named list of mungebit generators exposed by this package, invisibly.
}
\description{
All mungebit generators exported by the syberiaMungebits2 package.
}
