% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numeric_to_factor.R
\name{numeric_to_factor}
\alias{numeric_to_factor}
\title{numeric to factor helper function}
\usage{
numeric_to_factor(num, levs, na.to.missing = TRUE)
}
\arguments{
\item{num}{a numeric atomic vector. Will be restored to a factor variable.}

\item{levs}{a character vector of levels.}

\item{na.to.missing}{a logical. Whether to convert NAs to a "Missing" level.}
}
\description{
TODO: Make this faster
}
\keyword{internal}

