% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/orderer.R
\name{orderer}
\alias{orderer}
\title{Order a dataframe by a column}
\arguments{
\item{dataframe}{a data.frame to re-order}

\item{column_name}{the column by which to order}
}
\description{
Order a dataframe by a column
}

