% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{remove outliers based upon a z_score threshold z_score=> (x - mean(x) / sd(x)}
\arguments{
\item{x}{an numeric vector}

\item{threshold}{a z_score threshold from which the absolute value of a z_score above this threshold will be set to NA}
}
\description{
remove outliers based upon a z_score threshold z_score=> (x - mean(x) / sd(x)
}

