% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/renamer.R
\name{renamer}
\alias{renamer}
\title{Rename columns in a dataframe}
\arguments{
\item{dataframe}{a data.frame.}

\item{replacements}{a list of name replacements with keys old names and
respective values new names.}
}
\description{
Rename columns in a dataframe
}
\examples{
\dontrun{
  renamer(iris, list('Sepal.Length' = 'seplen', 'Sepal.Width' = 'sepwid'))
}
}

