% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_variable.R
\name{replace_variable}
\alias{replace_variable}
\title{Replaces a variable in the dataframe.}
\arguments{
\item{df}{dataframe. The dataframe to modify.}

\item{fun}{function. The function to run to convert the input into the new input.  Whichever argument is passed first to the function is the variable that will be replaced.}
}
\value{
Does not return anything, but modifies-in-place the dataframe to replace that variable.
}
\description{
Replaces a variable in the dataframe.
}
\examples{
# Repaces Sepal.Width to add 1 to each datapoint.
replace_variable()$run(iris, function(Sepal.Width) { Sepal.Width + 1 })

# Repaces Sepal.Length to add the corresponding Sepal.Width to each datapoint for Sepal.Length.
replace_variable()$run(iris, function(Sepal.Length, Sepal.Width) { Sepal.Width + Sepal.Length })

# Usage in a Syberia file
\dontrun{
list("Replace Sepal.Length with Sepal.Length + 1" = list(
 replace_variable,
 function(Sepal.Length) Sepal.Length + 1
))
}

}
\author{
Peter Hurford
}

