% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_categorical_variables.R
\name{restore_categorical_variables}
\alias{restore_categorical_variables}
\title{Restores categorical variables, allowing you to predict.}
\arguments{
\item{x.}{The column to restore.}
}
\description{
Restores categorical variables, allowing you to predict.
}

