% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_variables}
\alias{select_variables}
\title{Select variables in a dataframe.}
\arguments{
\item{dataframe}{a data.frame}

\item{cols}{an atomic vector. Drop all but these columns.}

\item{weak}{logical. Whether or not to add non-existent columns as NAs.}
}
\description{
Select variables in a dataframe.
}
\examples{
df <- iris; select_variables()$run(df, 1) # Select only first variable
df <- iris; select_variables()$run(df, c('Sepal.Length', 'Petal.Length'))
df <- iris; select_variables()$run(df, c(TRUE,TRUE,FALSE,FALSE,TRUE)) # Exclude cols 3 and 4
}

