% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.syberiaMungebits2.R
\docType{package}
\name{syberiaMungebits2}
\alias{syberiaMungebits2}
\alias{syberiaMungebits2-package}
\title{Common feature engineering preparation using mungebits2.}
\description{
Common feature engineering preparation using mungebits2.
}

