% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncator.R
\name{truncator}
\alias{truncator}
\title{Constrain a numeric column to the range [min,max] of the training set}
\arguments{
\item{x}{an atomic vector}

\item{digits}{how many decimal places you would like your numeric number to contain after its been truncated}
}
\description{
Constrain a numeric column to the range [min,max] of the training set
}
