% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_replacer.R
\name{value_replacer}
\alias{value_replacer}
\title{Replace arbitrary values in your data}
\arguments{
\item{x}{an atomic vector. The column to transform.}

\item{values_map}{a list. If named, the names get replaced w/ the values.
Otherwise, the list elements are lists of 2 elements, the first of
which are the values to be replaced and the second the replacement value.}
}
\value{
the replaced column
}
\description{
Replace arbitrary values in your data
}
\examples{
\dontrun{
# replace A, B, D with 1 and NA with 0.
value_replacer()$run(c("A", "B", NA, "D"), list(list(c("A","B","D"), 1), list(NA, 0)))
value_replacer()$run(c("A", "B", NA, "D"), list(A = 1, B = 1, D = 1, list(NA, 0)))
}
}
