% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_variable.R
\name{new_variable}
\alias{new_variable}
\title{Creates a new variable in the dataframe.}
\arguments{
\item{df}{dataframe. The dataframe to modify.}

\item{fun}{function. The function to run to convert old_variables to new_variables.
Also can take a non-function expression that will be converted to function() expression.}

\item{old_variables}{character. The names of the variables needed to make the new_variable.}

\item{new_variable}{character. The name of the desired new variable.}
}
\value{
Does not return anything, but modifies-in-place the dataframe to have that new variable.
}
\description{
Wrapper around the default multi_column_transformation mungebit to make new variables.
}
\examples{
new_variable()$run(iris, function(Sepal.Width, Sepal.Length) Sepal.Width*2 + Sepal.Length*2, 'Sepal.Perimeter')
#TODO: (RK) Fix broken example.
#new_variable()$run(iris, iris$Sepal.Width, 'Sepal.Width2')

\dontrun{
  # Usage in a Syberia file
  list("Make new variable" = list(
    new_variable,
    function(Sepal.Width, Sepal.Length) Sepal.Width*2 + Sepal.Length*2,
    'Sepal.Perimeter'
  ))
}
}

